+++ common/src/main/java/com/railwayteam/railways/content/schedule/RedstoneLinkInstruction.java
@@ -2,6 +2,7 @@ package com.railwayteam.railways.content.schedule;
 
 import com.google.common.collect.ImmutableList;
 import com.railwayteam.railways.Railways;
+import com.railwayteam.railways.annotation.event.MultiLoaderEvent;
 import com.railwayteam.railways.mixin.AccessorScheduleRuntime;
 import com.railwayteam.railways.mixin_interfaces.ICustomExecutableInstruction;
 import com.simibubi.create.AllBlocks;
@@ -36,6 +37,7 @@ public class RedstoneLinkInstruction extends ScheduleInstruction implements ICus
     public static WorldAttached<List<CustomRedstoneActor>> customActors =
         new WorldAttached<>($ -> new ArrayList<>());
 
+    @MultiLoaderEvent
     public static void tick(Level world) {
         List<CustomRedstoneActor> actors = customActors.get(world);
         for (Iterator<CustomRedstoneActor> actorIterator = actors.iterator(); actorIterator.hasNext();) {