+++ common/src/main/java/com/railwayteam/railways/registry/CRBogeyStyles.java
@@ -4,14 +4,16 @@ import com.google.common.collect.ImmutableList;
 import com.railwayteam.railways.Railways;
 import com.railwayteam.railways.api.bogeymenu.v0.BogeyMenuManager;
 import com.railwayteam.railways.api.bogeymenu.v0.entry.CategoryEntry;
-import com.railwayteam.railways.content.custom_bogeys.invisible.InvisibleBogeyRenderer;
-import com.railwayteam.railways.content.custom_bogeys.monobogey.InvisibleMonoBogeyBlock;
-import com.railwayteam.railways.content.custom_bogeys.monobogey.MonoBogeyRenderer;
+import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.large.*;
+import com.railwayteam.railways.content.custom_bogeys.special.invisible.InvisibleBogeyRenderer;
+import com.railwayteam.railways.content.custom_bogeys.special.monobogey.InvisibleMonoBogeyBlock;
+import com.railwayteam.railways.content.custom_bogeys.special.monobogey.MonoBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.narrow.NarrowDoubleScotchYokeBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.narrow.NarrowScotchYokeBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.narrow.NarrowSmallBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.HandcarBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.double_axle.*;
+import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.medium.*;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.single_axle.CoilspringBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.single_axle.LeafspringBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.single_axle.SingleaxleBogeyRenderer;
@@ -20,7 +22,7 @@ import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.standard.tr
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.wide.WideComicallyLargeScotchYokeBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.wide.WideDefaultBogeyRenderer;
 import com.railwayteam.railways.content.custom_bogeys.renderer.gauge.wide.WideScotchYokeBogeyRenderer;
-import com.railwayteam.railways.impl.bogeymenu.BogeyMenuManagerImpl;
+import com.railwayteam.railways.impl.bogeymenu.v0.BogeyMenuManagerImpl;
 import com.railwayteam.railways.registry.CRTrackMaterials.CRTrackType;
 import com.railwayteam.railways.util.Utils;
 import com.simibubi.create.AllBogeyStyles;
@@ -45,11 +47,17 @@ public class CRBogeyStyles {
     public static final String SINGLEAXLE_CYCLE_GROUP = "singleaxles";
     public static final String DOUBLEAXLE_CYCLE_GROUP = "doubleaxles";
     public static final String TRIPLEAXLE_CYCLE_GROUP = "tripleaxles";
+    public static final String QUADRUPLEAXLE_CYCLE_GROUP = "quadrupleaxles";
+    public static final String QUINTUPLEAXLE_CYCLE_GROUP = "quintupleaxles";
+    public static final String SEXTUPLEAXLE_CYCLE_GROUP = "sextupleaxles";
 
     public static final CategoryEntry STANDARD_CATEGORY = registerCategory(Create.ID, STANDARD_CYCLE_GROUP);
-    public static final CategoryEntry SINGLEAXLE_CATEGORY = registerCategory(Railways.MODID, SINGLEAXLE_CYCLE_GROUP);
-    public static final CategoryEntry DOUBLEAXLE_CATEGORY = registerCategory(Railways.MODID, DOUBLEAXLE_CYCLE_GROUP);
-    public static final CategoryEntry TRIPLEAXLE_CATEGORY = registerCategory(Railways.MODID, TRIPLEAXLE_CYCLE_GROUP);
+    public static final CategoryEntry SINGLEAXLE_CATEGORY = registerCategory(SINGLEAXLE_CYCLE_GROUP);
+    public static final CategoryEntry DOUBLEAXLE_CATEGORY = registerCategory(DOUBLEAXLE_CYCLE_GROUP);
+    public static final CategoryEntry TRIPLEAXLE_CATEGORY = registerCategory(TRIPLEAXLE_CYCLE_GROUP);
+    public static final CategoryEntry QUADRUPLEAXLE_CATEGORY = registerCategory(QUADRUPLEAXLE_CYCLE_GROUP);
+    public static final CategoryEntry QUINTUPLEAXLE_CATEGORY = registerCategory(QUINTUPLEAXLE_CYCLE_GROUP);
+    public static final CategoryEntry SEXTUPLEAXLE_CATEGORY = registerCategory(SEXTUPLEAXLE_CYCLE_GROUP);
 
     private static final Map<Pair<BogeyStyle, TrackType>, BogeyStyle> STYLES_FOR_GAUGES = new HashMap<>();
     private static final Map<BogeyStyle, BogeyStyle> STYLES_TO_STANDARD_GAUGE = new HashMap<>();
@@ -215,6 +223,76 @@ public class CRBogeyStyles {
         .soundType(AllSoundEvents.COGS.getId())
         .build();
 
+    // Medium
+    public static final BogeyStyle
+        MEDIUM_STANDARD = create("medium_standard", DOUBLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_standard"))
+            .size(BogeySizes.SMALL, () -> MediumStandardRenderer::new, CRBlocks.MEDIUM_BOGEY)
+            .build(),
+        MEDIUM_SINGLE_WHEEL = create("medium_single_wheel", SINGLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_single_wheel"))
+            .size(BogeySizes.SMALL, () -> MediumSingleWheelRenderer::new, CRBlocks.MEDIUM_BOGEY)
+            .build(),
+        MEDIUM_TRIPLE_WHEEL = create("medium_triple_wheel", TRIPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_triple_wheel"))
+            .size(BogeySizes.SMALL, () -> MediumTripleWheelRenderer::new, CRBlocks.MEDIUM_TRIPLE_WHEEL)
+            .build(),
+        MEDIUM_QUADRUPLE_WHEEL = create("medium_quadruple_wheel", QUADRUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_quadruple_wheel"))
+            .size(BogeySizes.SMALL, () -> MediumQuadrupleWheelRenderer::new, CRBlocks.MEDIUM_QUADRUPLE_WHEEL)
+            .build(),
+        MEDIUM_QUINTUPLE_WHEEL = create("medium_quintuple_wheel", QUINTUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_quintuple_wheel"))
+            .size(BogeySizes.SMALL, () -> MediumQuintupleWheelRenderer::new, CRBlocks.MEDIUM_QUINTUPLE_WHEEL)
+            .build(),
+        MEDIUM_2_0_2_TRAILING = create("medium_2_0_2_trailing", SINGLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_2_0_2_trailing"))
+            .size(BogeySizes.SMALL, () -> Medium202TrailingRenderer::new, CRBlocks.MEDIUM_2_0_2_TRAILING)
+            .build(),
+        MEDIUM_4_0_4_TRAILING = create("medium_4_0_4_trailing", DOUBLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_4_0_4_trailing"))
+            .size(BogeySizes.SMALL, () -> Medium404TrailingRenderer::new, CRBlocks.MEDIUM_4_0_4_TRAILING)
+            .build(),
+        MEDIUM_6_0_6_TRAILING = create("medium_6_0_6_trailing", TRIPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_6_0_6_trailing"))
+            .size(BogeySizes.SMALL, () -> Medium606TrailingRenderer::new, CRBlocks.MEDIUM_TRIPLE_WHEEL)
+            .build(),
+        MEDIUM_6_0_6_TENDER = create("medium_6_0_6_tender", TRIPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_6_0_6_tender"))
+            .size(BogeySizes.SMALL, () -> Medium606TenderRenderer::new, CRBlocks.MEDIUM_TRIPLE_WHEEL)
+            .build(),
+        MEDIUM_8_0_8_TENDER = create("medium_8_0_8_tender", QUADRUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_8_0_8_tender"))
+            .size(BogeySizes.SMALL, () -> Medium808TenderRenderer::new, CRBlocks.MEDIUM_QUADRUPLE_WHEEL)
+            .build(),
+        MEDIUM_10_0_10_TENDER = create("medium_10_0_10_tender", QUINTUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.medium_10_0_10_tender"))
+            .size(BogeySizes.SMALL, () -> Medium10010TenderRenderer::new, CRBlocks.MEDIUM_QUINTUPLE_WHEEL)
+            .build();
+
+    // Large
+    public static final BogeyStyle
+        LARGE_CREATE_STYLED_0_4_0 = create("large_create_style_0_4_0", DOUBLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.large_create_style_0_4_0"))
+            .size(BogeySizes.LARGE, () -> LargeCreateStyled040Renderer::new, CRBlocks.LARGE_CREATE_STYLE_0_4_0)
+            .build(),
+        LARGE_CREATE_STYLED_0_6_0 = create("large_create_style_0_6_0", TRIPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.large_create_style_0_6_0"))
+            .size(BogeySizes.LARGE, () -> LargeCreateStyled060Renderer::new, CRBlocks.LARGE_CREATE_STYLE_0_6_0)
+            .build(),
+        LARGE_CREATE_STYLED_0_8_0 = create("large_create_style_0_8_0", QUADRUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.large_create_style_0_8_0"))
+            .size(BogeySizes.LARGE, () -> LargeCreateStyled080Renderer::new, CRBlocks.LARGE_CREATE_STYLE_0_8_0)
+            .build(),
+        LARGE_CREATE_STYLED_0_10_0 = create("large_create_style_0_10_0", QUINTUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.large_create_style_0_10_0"))
+            .size(BogeySizes.LARGE, () -> LargeCreateStyled0100Renderer::new, CRBlocks.LARGE_CREATE_STYLE_0_10_0)
+            .build(),
+        LARGE_CREATE_STYLED_0_12_0 = create("large_create_style_0_12_0", SEXTUPLEAXLE_CYCLE_GROUP)
+            .displayName(Components.translatable("railways.bogeys.styles.large_create_style_0_12_0"))
+            .size(BogeySizes.LARGE, () -> LargeCreateStyled0120Renderer::new, CRBlocks.LARGE_CREATE_STYLE_0_12_0)
+            .build();
+
 
     public static AllBogeyStyles.BogeyStyleBuilder create(String name, String cycleGroup) {
         return create(Railways.asResource(name), Railways.asResource(cycleGroup))
@@ -227,6 +305,10 @@ public class CRBogeyStyles {
 
     public static AllBogeyStyles.BogeyStyleBuilder create(ResourceLocation name, ResourceLocation cycleGroup) {
         return new AllBogeyStyles.BogeyStyleBuilder(name, cycleGroup);
+    }
+
+    public static CategoryEntry registerCategory(String name) {
+        return registerCategory(Railways.MODID, name);
     }
 
     public static CategoryEntry registerCategory(String modid, String name) {
@@ -268,7 +350,7 @@ public class CRBogeyStyles {
 
         // Standard Category
         addToCategory(STANDARD_CATEGORY, INVISIBLE);
-        addToCategory(STANDARD_CATEGORY, WIDE_COMICALLY_LARGE, 18);
+        addToCategory(STANDARD_CATEGORY, WIDE_COMICALLY_LARGE, 17);
         addToCategory(STANDARD_CATEGORY, STANDARD);
         addToCategory(STANDARD_CATEGORY, NARROW_DOUBLE_SCOTCH);
 
@@ -276,6 +358,8 @@ public class CRBogeyStyles {
         addToCategory(SINGLEAXLE_CATEGORY, SINGLEAXLE);
         addToCategory(SINGLEAXLE_CATEGORY, COILSPRING);
         addToCategory(SINGLEAXLE_CATEGORY, LEAFSPRING);
+        addToCategory(SINGLEAXLE_CATEGORY, MEDIUM_SINGLE_WHEEL);
+        addToCategory(SINGLEAXLE_CATEGORY, MEDIUM_2_0_2_TRAILING);
 
         // Double Axle Category
         addToCategory(DOUBLEAXLE_CATEGORY, MODERN);
@@ -284,10 +368,29 @@ public class CRBogeyStyles {
         addToCategory(DOUBLEAXLE_CATEGORY, FREIGHT);
         addToCategory(DOUBLEAXLE_CATEGORY, PASSENGER);
         addToCategory(DOUBLEAXLE_CATEGORY, ARCHBAR);
+        addToCategory(DOUBLEAXLE_CATEGORY, MEDIUM_STANDARD);
+        addToCategory(DOUBLEAXLE_CATEGORY, MEDIUM_4_0_4_TRAILING);
+        addToCategory(DOUBLEAXLE_CATEGORY, LARGE_CREATE_STYLED_0_4_0);
 
         // Triple Axle Category
-        addToCategory(TRIPLEAXLE_CATEGORY, HEAVYWEIGHT);
-        addToCategory(TRIPLEAXLE_CATEGORY, RADIAL);
+        addToCategory(TRIPLEAXLE_CATEGORY, HEAVYWEIGHT, 20);
+        addToCategory(TRIPLEAXLE_CATEGORY, RADIAL, 20);
+        addToCategory(TRIPLEAXLE_CATEGORY, MEDIUM_6_0_6_TRAILING, 20);
+        addToCategory(TRIPLEAXLE_CATEGORY, MEDIUM_6_0_6_TENDER, 20);
+        addToCategory(TRIPLEAXLE_CATEGORY, LARGE_CREATE_STYLED_0_6_0, 20);
+
+        // Quadruple Axle Category
+        addToCategory(QUADRUPLEAXLE_CATEGORY, MEDIUM_QUADRUPLE_WHEEL, 19);
+        addToCategory(QUADRUPLEAXLE_CATEGORY, MEDIUM_8_0_8_TENDER, 19);
+        addToCategory(QUADRUPLEAXLE_CATEGORY, LARGE_CREATE_STYLED_0_8_0, 17);
+
+        // Quintuple Axle Category
+        addToCategory(QUINTUPLEAXLE_CATEGORY, MEDIUM_QUINTUPLE_WHEEL, 17);
+        addToCategory(QUINTUPLEAXLE_CATEGORY, MEDIUM_10_0_10_TENDER, 17);
+        addToCategory(QUINTUPLEAXLE_CATEGORY, LARGE_CREATE_STYLED_0_10_0, 15);
+
+        // Sextuple Axle Category
+        addToCategory(SEXTUPLEAXLE_CATEGORY, LARGE_CREATE_STYLED_0_12_0, 13);
 
         if (Utils.isDevEnv()) {
             CategoryEntry ALL_TEST_CATEGORY = registerCategory(Railways.MODID, "all_test");