+++ fabric/src/main/java/com/railwayteam/railways/registry/fabric/CRKeysImpl.java
@@ -7,4 +7,8 @@ public class CRKeysImpl {
     public static void registerKeyBinding(KeyMapping keyMapping) {
         KeyBindingHelper.registerKeyBinding(keyMapping);
     }
+
+    public static int getBoundCode(KeyMapping keyMapping) {
+        return KeyBindingHelper.getBoundKeyOf(keyMapping).getValue();
+    }
 }